housegen: fix CI failures (Windows C4244, profile section, binding docs)
- Cast std::atof (double) results to float in the setPerimeter binding (Windows
  /WX C4244).
- Move housegen's manifest declaration into the L5 section so its declared LAYER
  matches the section (profile-matrix contract).
- Document the new HouseRequest/HouseLayout script bindings in the user guide so
  the binding-contract check has no NEW gaps.

+++ src/modules/housegen/HouseGen.cpp
@@ -90,8 +90,8 @@ void HouseGen::expose(ssq::Table &table) {
                 const size_t comma = segment.find(',');
                 if (comma != std::string::npos) {
                     HousePolygonPoint p;
-                    p.x = std::atof(segment.substr(0, comma).c_str());
-                    p.y = std::atof(segment.substr(comma + 1).c_str());
+                    p.x = static_cast<float>(std::atof(segment.substr(0, comma).c_str()));
+                    p.y = static_cast<float>(std::atof(segment.substr(comma + 1).c_str()));
                     r->perimeter.push_back(p);
                 }
                 segment.clear();
